requirements.txt
PyQt5
pandas
openpyxl
faker
Pillow
pyinstaller


